+++ gatsby-browser.tsx
@@ -4,6 +4,7 @@ import './src/styles/global.css'
 import HandbookLayout from './src/templates/Handbook'
 import Product from './src/templates/Product'
 import Job from './src/templates/Job'
+import Posts from './src/components/Edition/Posts'
 import { Provider as ToastProvider } from './src/context/toast'
 import { RouteUpdateArgs } from 'gatsby'
 import { UserProvider } from './src/hooks/useUser'
@@ -56,6 +57,8 @@ export const wrapPageElement = ({ element, props }) => {
         <Product {...props} />
     ) : /^careers\//.test(slug) ? (
         <Job {...props} />
+    ) : /^blog|^tutorials/.test(slug) ? (
+        <Posts {...props}>{element}</Posts>
     ) : (
         element
     )
