+++ src/templates/tutorials/Tutorial.tsx
@@ -44,35 +44,6 @@ export const ViewButton = ({ title, view, setView }) => {
 
 const A = (props) => <Link {...props} className="text-red hover:text-red font-semibold" />
 
-const TutorialSidebar = ({ contributors, location, title, categories }) => {
-    return (
-        <>
-            {contributors?.length > 0 && (
-                <SidebarSection>
-                    <Contributors
-                        contributors={contributors.map((contributor) => ({
-                            ...contributor,
-                        }))}
-                    />
-                </SidebarSection>
-            )}
-            <SidebarSection title="Share">
-                <ShareLinks title={title} href={location.href} />
-            </SidebarSection>
-            {categories?.length > 0 && (
-                <SidebarSection title="Filed under...">
-                    <Topics
-                        topics={categories?.map((category) => ({
-                            name: category,
-                            url: `/tutorials/categories/${slugify(category, { lower: true })}`,
-                        }))}
-                    />
-                </SidebarSection>
-            )}
-        </>
-    )
-}
-
 export default function Tutorial({ data, pageContext: { tableOfContents, menu }, location }) {
     const { pageData } = data
     const { body, excerpt, fields } = pageData
@@ -98,65 +69,39 @@ export default function Tutorial({ data, pageContext: { tableOfContents, menu },
     const [view, setView] = useState('Article')
 
     return (
-        <Layout parent={communityMenu} activeInternalMenu={communityMenu.children[1]}>
+        <>
             <SEO
                 title={title + ' - PostHog'}
                 description={description || excerpt}
                 article
                 image={`/og-images/${fields.slug.replace(/\//g, '')}.jpeg`}
             />
-            <PostLayout
-                isMenuItemActive={({ url }) =>
-                    url === `/tutorials/categories/${slugify(categories[0], { lower: true })}`
-                }
-                questions={
-                    <div id="squeak-questions" className="pb-8">
-                        <CommunityQuestions />
-                    </div>
-                }
-                body={body}
+            <Intro
+                contributors={contributors}
                 featuredImage={featuredImage}
-                featuredVideo={featuredVideo}
-                filePath={filePath}
-                tableOfContents={tableOfContents}
                 title={title}
-                menu={menu}
-                sidebar={
-                    <TutorialSidebar
-                        contributors={contributors}
-                        location={location}
-                        title={title}
-                        categories={categories}
-                    />
-                }
-            >
-                <Intro
-                    contributors={contributors}
-                    featuredImage={featuredImage}
-                    title={title}
-                    featuredImageType="full"
-                    titlePosition="top"
-                    date={date}
-                />
+                featuredImageType="full"
+                titlePosition="top"
+                date={date}
+            />
 
-                {featuredVideo && (
-                    <div className="mb-6 flex space-x-2">
-                        <ViewButton view={view} title="Article" setView={setView} />
-                        <ViewButton view={view} title="Video" setView={setView} />
-                    </div>
-                )}
-                {view === 'Article' && breakpoints.md && <MobileSidebar tableOfContents={tableOfContents} />}
-                {view === 'Article' ? (
-                    <div className="article-content">
-                        <MDXProvider components={components}>
-                            <MDXRenderer>{body}</MDXRenderer>
-                        </MDXProvider>
-                    </div>
-                ) : (
-                    <iframe src={featuredVideo} />
-                )}
-            </PostLayout>
-        </Layout>
+            {featuredVideo && (
+                <div className="mb-6 flex space-x-2">
+                    <ViewButton view={view} title="Article" setView={setView} />
+                    <ViewButton view={view} title="Video" setView={setView} />
+                </div>
+            )}
+            {view === 'Article' && breakpoints.md && <MobileSidebar tableOfContents={tableOfContents} />}
+            {view === 'Article' ? (
+                <div className="article-content">
+                    <MDXProvider components={components}>
+                        <MDXRenderer>{body}</MDXRenderer>
+                    </MDXProvider>
+                </div>
+            ) : (
+                <iframe src={featuredVideo} />
+            )}
+        </>
     )
 }
